src/kafka/requirements.txt
python-dotenv
kafka-python
faker
jsonschema
certifi 
